refactor(Options): Move options to a separate file
feat(Options): Added verbose option.
chore(Options): Renamed `recurse` to `recursive`.
